+++ src/components/features/protocol/components/authenticated-by.tsx
