+++ src/components/features/protocol/components/copy-paste.tsx
