+++ src/components/features/protocol/context.ts
@@ -1,0 +1,9 @@
+import { createContext, useContext } from 'react';
+
+export type ProtocolContextType = {
+  qrCode?: string;
+};
+
+export const ProtocolContext = createContext<ProtocolContextType>(null);
+
+export const useProtocolContext = () => useContext(ProtocolContext);
