+++ src/components/features/protocol/credentials/create/components/claim-image-field.tsx
