+++ src/components/features/protocol/credentials/show/components/credential-title-and-image.tsx
