+++ src/components/features/protocol/credentials/show/components/image-view.tsx
