+++ src/components/features/protocol/credentials/show/components/invalid-status-box.tsx
