+++ src/components/features/protocol/credentials/show/credential-show.tsx
