+++ src/components/features/protocol/data-models/show/components/table-schema.tsx
