+++ src/pages/model/[id]/playground.tsx
@@ -1,7 +1,7 @@
 import dynamic from 'next/dynamic';
 
-import { PlaygroundTab } from '@/components/templates/protocol/data-models/show/components/playground-tab';
-import { DataModelLayout } from '@/components/templates/protocol/data-models/show/layout';
+import { PlaygroundTab } from '@/components/features/protocol/data-models/show/components/playground-tab';
+import { DataModelLayout } from '@/components/features/protocol/data-models/show/layout';
 
 export default function ProtocolDataModelPlayground() {
   return <PlaygroundTab />;
