+++ src/pages/model/[id]/recipients.tsx
@@ -3,13 +3,13 @@ import { useRouter } from 'next/router';
 
 import { IColumnGrid } from '@/components/organisms/data-grid/data-grid';
 import { DataModelsTab } from '@/components/templates/explore';
-import { DataModelLayout } from '@/components/templates/protocol/data-models/show/layout';
+import { DataModelLayout } from '@/components/features/protocol/data-models/show/layout';
 import { query } from '@/constants/queries';
 
 const GridViewTab = dynamic(
   () =>
     import(
-      '@/components/templates/protocol/data-models/show/components/grid-view-tab'
+      '@/components/features/protocol/data-models/show/components/grid-view-tab'
     ),
   { ssr: false }
 );
